Install latest kernel / kernel-headers
In order to satisfy both vbox guest and patched lustre, we need to build boxes for each supported kernel.

We can do this by both pinning the kernel headers (so they don't upgrade on vbox guest install), and building a box for each kernel version.

Upgrade to the latest kernel. A new version of the box will need to be created each time the kernel revs and Lustre intends to support it.

Signed-off-by: Joe Grund <joe.grund@intel.com>