Add template to create a box that will be used to build lustre on top of (#18)
mellanox

The process of building lustre on top of mellanox drivers and
distributing the rpms is time consuming. Vagrant can do a lot of the
heavy lifting and provisioning scripts can be used to automate the
build process. Create templates that will be used to build out a box
that has all dependencies and the appropriate lustre kernel ready to go
so that anyone can create the rpms without the hassle.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>